% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{do_initial_transform}
\alias{do_initial_transform}
\title{Do an initial transformation of time series data.}
\usage{
do_initial_transform(
  y,
  transformation = c("none", "log", "sqrt", "box-cox"),
  transform_offset = 0.5,
  bc_lambda
)
}
\arguments{
\item{y}{a univariate time series or numeric vector.}

\item{transformation}{character specifying transformation type:
"box-cox", "sqrt", "log", "forecast-box-cox", or "none".}

\item{transform_offset}{numeric offset used before the Box-Cox, sqrt, and log
transformations; the offset is added to all observations before
transforming.  Default value of 0.5 allows us to use the Box-Cox and log
transforms (which require positive inputs) in case of observations of 0,
and also ensures that the reverse transformed values will always be at
least -0.5, so that they round up to non-negative values.}

\item{bc_lambda}{(required if transformation is "box-cox").  Parameter for
the Box-Cox transformation.  bc_lambda should be the result of a
call to car::powerTransform(y + transform_offset, family = "bcPower")}
}
\value{
a transformed object of the same class as y
}
\description{
Do an initial transformation of time series data.
}
