% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_baseline.R
\name{is.quantile_baseline}
\alias{is.quantile_baseline}
\title{Check if object is of class quantile_baseline}
\usage{
is.quantile_baseline(object)
}
\arguments{
\item{object}{an object that may be a quantile_baseline object}
}
\value{
boolean; whether object is inherits quantile_baseline class
}
\description{
Check if object is of class quantile_baseline
}
