% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict.simple_ts}
\alias{predict.simple_ts}
\title{Generate predictions from a model fit with class simple_ts}
\usage{
\method{predict}{simple_ts}(simple_ts_fit, newdata = NULL, horizon = 1, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{newdata}{numeric vector of new data to simulate forward from}

\item{horizon}{number of time steps forwards to simulate}

\item{nsim}{number of sample trajectories to simulate}

\item{seed}{either \code{NULL} or an integer that will be used in a call to
\code{set.seed} before simulating the response vectors.  If set, the value is
saved as the "seed" attribute of the returned value.  The default, \code{NULL},
will not change the random generator state, and return \code{.Random.seed}
as the "seed" attribute}

\item{...}{other arguments passed on to model-specific predict methods}

\item{object}{a model fit of class "simple_ts", as returned by fit_simple_ts}
}
\value{
an nsim by horizon matrix with simulated values
}
\description{
This function linearly interpolates missing values on the interior of the
time series, which is necessary for some model fits with moving average
components.  It also handles transformations and possible seasonal
differencing that may have been done in the fit_simple_ts function.
}
