% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplets-package.R
\docType{package}
\name{simplets}
\alias{simplets}
\alias{simplets-package}
\title{simplets - Simple models for time series forecasting}
\author{
Evan L. Ray \href{mailto:elray@umass.edu}{elray@umass.edu}
}
